Hello-spring/src/main/java/Hello/Hellospring/Repository/MemoryMemberRepository.java
package Hello.Hellospring.Repository;

import Hello.Hellospring.Domain.Member;
import org.springframework.stereotype.Repository;

import java.util.*;

@Repository
public class MemoryMemberRepository implements MemberRepository {

    private static Map<Long, Member> store = new HashMap<>();
    private static long sequence = 0L;

    @Override // @Override: 부모클래스의 기능 변경
    public Member save(Member member) {
        member.setId(++sequence);
        store.put(member.getId(), member);
        return member;
    }
    @Override
    public Optional<Member> findById(Long id) {
        return Optional.ofNullable(store.get(id));
        // ofNullable: value가 null인 경우에도 반환합니다.
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    @Override
    public Optional<Member> findByName(String name) {
        return store.values().stream()
                // ->(람다식): 식별자 없이 실행가능한 함
                .filter(member -> member.getName().equals(name))
                .findAny();
    }
    public void clearStore() {
        store.clear(); // store 데이터를 초기화 시킴
    }
}
